Introduce actionlogs plugin group
