+++ administrator/components/com_userlogs/config.xml
@@ -1,5 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <config>
+	<fieldset name="userlogs" label="COM_USERLOGS_OPTIONS" addfieldpath="/administrator/components/com_userlogs/models/fields">
+		<field
+				name="ip_logging"
+				type="radio"
+				label="COM_USERLOGS_IP_LOGGING"
+				description="COM_USERLOGS_IP_LOGGING_DESC"
+				class="btn-group btn-group-yesno"
+				default="0"
+		>
+			<option value="1">JYES</option>
+			<option value="0">JNO</option>
+		</field>
+		<field
+				name="loggable_extensions"
+				type="logtype"
+				label="COM_USERLOGS_LOG_EXTENSIONS"
+				description="COM_USERLOGS_LOG_EXTENSIONS_DESC"
+				multiple="true"
+				default="com_banners,com_cache,com_categories,com_config,com_contact,com_content,com_installer,com_media,com_menus,com_messages,com_modules,com_newsfeeds,com_plugins,com_redirect,com_tags,com_templates,com_users"
+		/>
+	</fieldset>
 	<fieldset
 		name="permissions"
 		label="JCONFIG_PERMISSIONS_LABEL"
