+++ administrator/language/en-GB/en-GB.com_userlogs.ini
@@ -8,6 +8,9 @@ COM_USERLOGS_ACTION_VIEWLOGS="View logs"
 COM_USERLOGS_ACTION_VIEWLOGS_DESC="Allows users in the group to view the logs."
 COM_USERLOGS_CONFIGURATION="Userlogs: Options"
 COM_USERLOGS_DATE="Date"
+COM_USERLOGS_DISABLED="Disabled"
+COM_USERLOGS_EMAIL_DESC="This is the latest action done by a user on your website"
+COM_USERLOGS_EMAIL_SUBJECT="Latest User Actions"
 COM_USERLOGS_EXPORT_ALL_CSV="Export All as CSV"
 COM_USERLOGS_EXPORT_CSV="Export Selected as CSV"
 COM_USERLOGS_EXTENSION="Extension"
@@ -17,10 +20,17 @@ COM_USERLOGS_EXTENSION_FILTER_DESC="Search in the User Action logs by extension"
 COM_USERLOGS_IP_ADDRESS="IP Address"
 COM_USERLOGS_IP_ADDRESS_ASC="IP Address ascending"
 COM_USERLOGS_IP_ADDRESS_DESC="IP Address descending"
+COM_USERLOGS_IP_LOGGING="Enable IP logging"
+COM_USERLOGS_IP_LOGGING_DESC="Enable or disable logging the IP address of the users"
+COM_USERLOGS_LOG_EXTENSIONS="Select events to be logged"
+COM_USERLOGS_LOG_EXTENSIONS_DESC="Select events to be logged"
 COM_USERLOGS_MANAGER_USERLOGS="User Actions Log"
 COM_USERLOGS_MESSAGE="Message"
 COM_USERLOGS_MESSAGE_ASC="Message ascending"
 COM_USERLOGS_MESSAGE_DESC="Message descending"
+COM_USERLOGS_NAME="Name"
+COM_USERLOGS_NAME_ASC="Name ascending"
+COM_USERLOGS_NAME_DESC="Name descending"
 COM_USERLOGS_N_ITEMS_DELETED="%s logs deleted."
 COM_USERLOGS_N_ITEMS_DELETED_1="%s log deleted."
 COM_USERLOGS_NO_LOGS_TO_EXPORT="There are no User Action logs to export."
@@ -34,13 +44,11 @@ COM_USERLOGS_OPTION_RANGE_PAST_WEEK="In the last week"
 COM_USERLOGS_OPTION_RANGE_PAST_YEAR="In the last year"
 COM_USERLOGS_OPTION_RANGE_POST_YEAR="More than a year ago"
 COM_USERLOGS_OPTION_RANGE_TODAY="Today"
+COM_USERLOGS_OPTIONS="Options"
 COM_USERLOGS_PURGE_CONFIRM="Are you sure want to delete all User Action logs?"
 COM_USERLOGS_PURGE_FAIL="Failed to delete all User Action logs."
 COM_USERLOGS_PURGE_SUCCESS="All User Action logs have been deleted."
-COM_USERLOGS_TOOLBAR_PURGE="Purge"
 COM_USERLOGS_SELECT_EXTENSION="- Select Extension -"
 COM_USERLOGS_SELECT_USER="- Select User -"
-COM_USERLOGS_NAME="Name"
-COM_USERLOGS_NAME_ASC="Name ascending"
-COM_USERLOGS_NAME_DESC="Name descending"
+COM_USERLOGS_TOOLBAR_PURGE="Purge"
 COM_USERLOGS_XML_DESCRIPTION="The component that views the log of the actions of different users."
