+++ administrator/language/en-GB/en-GB.plg_actionlogs_joomla.ini
@@ -1,0 +1,38 @@
+; Joomla! Project
+; Copyright (C) 2005 - 2018 Open Source Matters. All rights reserved.
+; License GNU General Public License version 2 or later; see LICENSE.txt, see LICENSE.php
+; Note : All ini files need to be saved as UTF-8
+
+PLG_ACTIONLOGS_JOOMLA="Action Logs - Joomla"
+PLG_ACTIONLOGS_JOOMLA_XML_DESCRIPTION="Record the actions of users on the site for Joomla core extensions so they can be reviewed if required."
+; Content types title
+PLG_SYSTEM_USERLOGS_TYPE_ACCESS_LEVEL="access level"
+PLG_SYSTEM_USERLOGS_TYPE_ARTICLE="article"
+PLG_SYSTEM_USERLOGS_TYPE_BANNER="banner"
+PLG_SYSTEM_USERLOGS_TYPE_BANNER_CLIENT="banner client"
+PLG_SYSTEM_USERLOGS_TYPE_CATEGORY="category"
+PLG_SYSTEM_USERLOGS_TYPE_COMPONENT="component"
+PLG_SYSTEM_USERLOGS_TYPE_COMPONENT_CONFIG="Component Configuration"
+PLG_SYSTEM_USERLOGS_TYPE_CONTACT="contact"
+PLG_SYSTEM_USERLOGS_TYPE_LIBRARY="library"
+PLG_SYSTEM_USERLOGS_TYPE_LINK="link redirect"
+PLG_SYSTEM_USERLOGS_TYPE_LINK_REDIRECT="link redirect"
+PLG_SYSTEM_USERLOGS_TYPE_MEDIA="media"
+PLG_SYSTEM_USERLOGS_TYPE_MENU="menu"
+PLG_SYSTEM_USERLOGS_TYPE_MENU_ITEM="menu item"
+PLG_SYSTEM_USERLOGS_TYPE_MODULE="module"
+PLG_SYSTEM_USERLOGS_TYPE_NEWSFEED="newsfeed"
+PLG_SYSTEM_USERLOGS_TYPE_PLUGIN="plugin"
+PLG_SYSTEM_USERLOGS_TYPE_STYLE="template style"
+PLG_SYSTEM_USERLOGS_TYPE_TAG="tag"
+PLG_SYSTEM_USERLOGS_TYPE_TEMPLATE="template"
+PLG_SYSTEM_USERLOGS_TYPE_TEMPLATE_STYPE="template style"
+PLG_SYSTEM_USERLOGS_TYPE_USER="user"
+PLG_SYSTEM_USERLOGS_TYPE_USER_GROUP="user group"
+PLG_SYSTEM_USERLOGS_TYPE_USER_NOTE="user note"
+; Component
+PLG_SYSTEM_USERLOGS_COMPONENT_CONFIG_UPDATED="User <a href="{accountlink}">{username}</a> change settings of the component {extension_name}"
+; Extensions
+PLG_SYSTEM_USERLOGS_EXTENSION_INSTALLED="User <a href="_{accountlink}">{username}</a> installed the {type} {extension_name}"
+PLG_SYSTEM_USERLOGS_EXTENSION_UNINSTALLED="User <a href="{accountlink}">{username}</a> uninstalled the {type} {extension_name}"
+PLG_SYSTEM_USERLOGS_EXTENSION_UPDATED="User <a href="_{accountlink}">{username}</a> updated the {type} {extension_name}"
