+++ administrator/language/en-GB/en-GB.plg_system_userlogs.ini
@@ -4,45 +4,13 @@
 ; Note : All ini files need to be saved as UTF-8
 
 PLG_SYSTEM_USERLOGS="System - User Actions Log"
-PLG_SYSTEM_USERLOGS_DISABLED="Disabled"
-PLG_SYSTEM_USERLOGS_EMAIL_DESC="This is the latest action done by a user on your website"
-PLG_SYSTEM_USERLOGS_EMAIL_SUBJECT="Latest User Actions"
 PLG_SYSTEM_USERLOGS_EXTENSIONS_NOTIFICATIONS="Select events to be notified for"
 PLG_SYSTEM_USERLOGS_EXTENSIONS_NOTIFICATIONS_DESC="Select events to be sent to your email as notifications"
-PLG_SYSTEM_USERLOGS_IP_LOGGING="Enable IP logging"
-PLG_SYSTEM_USERLOGS_IP_LOGGING_DESC="Enable or disable logging the IP address of the users"
 PLG_SYSTEM_USERLOGS_LOG_DELETE_PERIOD="Days to delete logs after"
 PLG_SYSTEM_USERLOGS_LOG_DELETE_PERIOD_DESC="Enter how many days logs should be kept before they are deleted. Enter 0 if you don't want to delete the logs."
-PLG_SYSTEM_USERLOGS_LOG_EXTENSIONS="Select events to be logged"
-PLG_SYSTEM_USERLOGS_LOG_EXTENSIONS_DESC="Select events to be logged"
 PLG_SYSTEM_USERLOGS_NOTIFICATIONS="Send notifications for User Actions Log"
 PLG_SYSTEM_USERLOGS_NOTIFICATIONS_DESC="Send a notifications of users' actions log to your email"
 PLG_SYSTEM_USERLOGS_XML_DESCRIPTION="Record the actions of users on the site so they can be reviewed if required."
-; Content type title
-PLG_SYSTEM_USERLOGS_TYPE_ACCESS_LEVEL="access level"
-PLG_SYSTEM_USERLOGS_TYPE_ARTICLE="article"
-PLG_SYSTEM_USERLOGS_TYPE_BANNER="banner"
-PLG_SYSTEM_USERLOGS_TYPE_BANNER_CLIENT="banner client"
-PLG_SYSTEM_USERLOGS_TYPE_CATEGORY="category"
-PLG_SYSTEM_USERLOGS_TYPE_COMPONENT="component"
-PLG_SYSTEM_USERLOGS_TYPE_COMPONENT_CONFIG="Component Configuration"
-PLG_SYSTEM_USERLOGS_TYPE_CONTACT="contact"
-PLG_SYSTEM_USERLOGS_TYPE_LIBRARY="library"
-PLG_SYSTEM_USERLOGS_TYPE_LINK="link redirect"
-PLG_SYSTEM_USERLOGS_TYPE_LINK_REDIRECT="link redirect"
-PLG_SYSTEM_USERLOGS_TYPE_MEDIA="media"
-PLG_SYSTEM_USERLOGS_TYPE_MENU="menu"
-PLG_SYSTEM_USERLOGS_TYPE_MENU_ITEM="menu item"
-PLG_SYSTEM_USERLOGS_TYPE_MODULE="module"
-PLG_SYSTEM_USERLOGS_TYPE_NEWSFEED="newsfeed"
-PLG_SYSTEM_USERLOGS_TYPE_PLUGIN="plugin"
-PLG_SYSTEM_USERLOGS_TYPE_STYLE="template style"
-PLG_SYSTEM_USERLOGS_TYPE_TAG="tag"
-PLG_SYSTEM_USERLOGS_TYPE_TEMPLATE="template"
-PLG_SYSTEM_USERLOGS_TYPE_TEMPLATE_STYPE="template style"
-PLG_SYSTEM_USERLOGS_TYPE_USER="user"
-PLG_SYSTEM_USERLOGS_TYPE_USER_GROUP="user group"
-PLG_SYSTEM_USERLOGS_TYPE_USER_NOTE="user note"
 ; Common content type log messages
 PLG_SYSTEM_USERLOGS_CONTENT_ADDED="User <a href="{accountlink}">{username}</a> added new {type} <a href="{itemlink}">{title}</a>"
 PLG_SYSTEM_USERLOGS_CONTENT_UPDATED="User <a href="{accountlink}">{username}</a> updated the {type} <a href="{itemlink}">{title}</a>"
@@ -50,9 +18,3 @@ PLG_SYSTEM_USERLOGS_CONTENT_PUBLISHED="User <a href="{accountlink}">{username}</
 PLG_SYSTEM_USERLOGS_CONTENT_UNPUBLISHED="User <a href="{accountlink}">{username}</a> unpublished the {type} <a href="{itemlink}">{title}</a>"
 PLG_SYSTEM_USERLOGS_CONTENT_TRASHED="User <a href="{accountlink}">{username}</a> trashed the {type} <a href="{itemlink}">{title}</a>"
 PLG_SYSTEM_USERLOGS_CONTENT_DELETED="User <a href="{accountlink}">{username}</a> deleted the {type} {title}"
-; Component
-PLG_SYSTEM_USERLOGS_COMPONENT_CONFIG_UPDATED="User <a href="{accountlink}">{username}</a> change settings of the component {extension_name}"
-; Extensions
-PLG_SYSTEM_USERLOGS_EXTENSION_INSTALLED="User <a href="_{accountlink}">{username}</a> installed the {type} {extension_name}"
-PLG_SYSTEM_USERLOGS_EXTENSION_UNINSTALLED="User <a href="{accountlink}">{username}</a> uninstalled the {type} {extension_name}"
-PLG_SYSTEM_USERLOGS_EXTENSION_UPDATED="User <a href="_{accountlink}">{username}</a> updated the {type} {extension_name}"
