+++ plugins/system/actionlogs/joomla/joomla.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<extension version="3.9" type="plugin" group="actionlogs" method="upgrade">
+	<name>PLG_ACTIONLOGS_Joomla</name>
+	<author>Joomla! Project</author>
+	<creationDate>May 2018</creationDate>
+	<copyright>Copyright (C) 2005 - 2018 Open Source Matters. All rights reserved.</copyright>
+	<license>GNU General Public License version 2 or later; see LICENSE.txt</license>
+	<authorEmail>admin@joomla.org</authorEmail>
+	<authorUrl>www.joomla.org</authorUrl>
+	<version>__DEPLOY_VERSION__</version>
+	<description>PLG_ACTIONLOGS_JOOMLA_XML_DESCRIPTION</description>
+	<files>
+		<filename plugin="joomla">joomla.php</filename>		
+	</files>
+	<config />
+</extension>
