--- plugins/system/userlogs/forms/userlogs.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form>
-	<fields name="params">
-		<fieldset name="settings" addfieldpath="/plugins/system/userlogs/field">
-			<field
-				name="logs_notification_option"
-				type="radio"
-				label="PLG_SYSTEM_USERLOGS_NOTIFICATIONS"
-				description="PLG_SYSTEM_USERLOGS_NOTIFICATIONS_DESC"
-				class="btn-group btn-group-yesno"
-				default="0"
-				required="true"
-				>
-				<option value="1">JYES</option>
-				<option value="0">JNO</option>
-			</field>
-			<field
-				name="logs_notification_extensions"
-				type="logtype"
-				label="PLG_SYSTEM_USERLOGS_EXTENSIONS_NOTIFICATIONS"
-				description="PLG_SYSTEM_USERLOGS_EXTENSIONS_NOTIFICATIONS_DESC"
-				multiple="true"
-			/>
-		</fieldset>
-	</fields>
-</form>
