+++ plugins/system/userlogs/userlogs.xml
@@ -12,37 +12,16 @@
 	<files>
 		<filename plugin="userlogs">userlogs.php</filename>
 		<folder>forms</folder>
-		<folder>field</folder>
-		<folder>layouts</folder>
 	</files>
 	<config>
 		<fields name="params">
-			<fieldset name="basic" addfieldpath="/plugins/system/userlogs/models/field">
+			<fieldset name="basic">
 				<field
 					name="logDeletePeriod"
 					type="number"
 					label="PLG_SYSTEM_USERLOGS_LOG_DELETE_PERIOD"
 					description="PLG_SYSTEM_USERLOGS_LOG_DELETE_PERIOD_DESC"
 					default="0"
-				/>
-				<field
-					name="ip_logging"
-					type="radio"
-					label="PLG_SYSTEM_USERLOGS_IP_LOGGING"
-					description="PLG_SYSTEM_USERLOGS_IP_LOGGING_DESC"
-					class="btn-group btn-group-yesno"
-					default="0"
-					>
-					<option value="1">JYES</option>
-					<option value="0">JNO</option>
-				</field>
-				<field
-					name="loggable_extensions"
-					type="logtype"
-					label="PLG_SYSTEM_USERLOGS_LOG_EXTENSIONS"
-					description="PLG_SYSTEM_USERLOGS_LOG_EXTENSIONS_DESC"
-					multiple="true"
-					default="com_banners,com_cache,com_categories,com_config,com_contact,com_content,com_installer,com_media,com_menus,com_messages,com_modules,com_newsfeeds,com_plugins,com_redirect,com_tags,com_templates,com_users"
 				/>
 				<field
 						name="lastrun"
